requirements.txt
aiogram==3.0.0b7
python-dotenv==0.20.0
aiohttp~=3.8.4
pydantic~=1.10.7
SQLAlchemy~=2.0.7
alembic
asyncpg
python-dateutil